requirements.txt
torch
torchvision
ultralytics
numpy
Pillow
transformers
fastapi
uvicorn
